+++ src/constants.ts
@@ -1,14 +1,14 @@
 export enum Inputs {
-    Key = "key",
-    Path = "path",
-    RestoreKeys = "restore-keys",
-    UploadChunkSize = "upload-chunk-size",
-    FailOnCacheMiss = "fail-on-cache-miss",
-    SaveOnAnyFailure = "save-on-any-failure"
+    Key = "key", // Input for cache, restore, save action
+    Path = "path", // Input for cache, restore, save action
+    RestoreKeys = "restore-keys", // Input for cache, restore action
+    UploadChunkSize = "upload-chunk-size" // Input for cache, save action
 }
 
 export enum Outputs {
-    CacheHit = "cache-hit"
+    CacheHit = "cache-hit", // Output from cache, restore action
+    CachePrimaryKey = "cache-primary-key", // Output from restore action
+    CacheRestoreKey = "cache-restore-key" // Output from restore action
 }
 
 export enum State {
@@ -20,10 +20,6 @@ export enum Events {
     Key = "GITHUB_EVENT_NAME",
     Push = "push",
     PullRequest = "pull_request"
-}
-
-export enum Variables {
-    SaveCacheOnAnyFailure = "SAVE_CACHE_ON_ANY_FAILURE"
 }
 
 export const RefKey = "GITHUB_REF";